requirements.txt
autopep8==1.4.4
beautifulsoup4==4.8.0
bs4==0.0.1
Click==7.0
Django==2.2.6
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
html5lib==1.0.1
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.3
pycodestyle==2.5.0
pyflakes==2.1.1
pymongo==3.9.0
python-dotenv==0.10.3
pytz==2019.3
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
webencodings==0.5.1
Werkzeug==0.15.6
yapf==0.28.0
